feat(apollo-react): add pre-hook actions for chat service [PLT-000] (#4058)
## Description

<!--
- Test portal-shell changes using the [Chrome
extension](https://github.com/UiPath/apollo-design-system/tree/master/tools/portal-devtools)
- Check if you need to update storybook
- If the package size increased drastically, add the difference here
- Add Jira ticket in the title
-->

### Why/Reason for changes
In order for our team to support long running workflows for
conversational agents (V4 chat experience), we need users to have to
select an agent successfully before creating a new chat. The ideal
workflow would be:
1. A user clicks "New Chat", and will be given a popup modal which
allows them to search/select one agent from all that are in their
tenant, and then confirm
2. Once the user clicks confirm, we handle any specific business logic
on our end and upon closing the modal, either do nothing or open the new
chat as the user wanted

### How/What are the changes doing
In order to support it, we're allowing consumers of the chat service to
add pre-hooks to exposed actions (only header actions for the moment)
and the pre-hooks are promises that resolve with a boolean; true meaning
proceed, and false meaning do not perform the action the user wanted to
do. **This allows control to be given to the consumers for wether the
action should take place or not, or decorate with their own custom
experiences as needed.**

### Demo
Below is a small demo in the sandbox that resembles how we want it to
work to support V4 chat experience

https://github.com/user-attachments/assets/f3cfe1c2-010e-4d57-bc7c-7a0f9fefbdd7

## Risk to rolling out

- Add risks of rolling out, including any additions/removals of feature
flags or APIs

<!-- JIRA START -->
## Jira
No Jira tickets found in PR title.
<!-- JIRA END -->

---------

Co-authored-by: Shannon Suhendra <shannon@suhendra.com>